remove latex_error only if it exists

+++ tex2pdf.sh
@@ -8,7 +8,9 @@ if grep -q "!" "temp.log"; then
   cat temp.log > latex_error
   awk '/\!/{flag=1} flag' temp.log | head
 else
-  rm latex_error
+  if [ -f latex_error ]; then
+    rm latex_error
+  fi
 fi
 
 ls | grep -e 'dvi$' -e '.aux$' -e '.nav$' -e '.log$' -e '.out$' -e '.snm$' -e '.toc$' | xargs rm
